+++ packages/nimble-components/src/table-column/base/models/column-internals.ts
@@ -1,6 +1,10 @@
 import { uniqueId } from '@microsoft/fast-web-utilities';
 import { ViewTemplate, observable } from '@microsoft/fast-element';
-import { TableColumnSortDirection, TableFieldName } from '../../../table/types';
+import {
+    TableColumnAlignment,
+    TableColumnSortDirection,
+    TableFieldName
+} from '../../../table/types';
 import type { TableCell } from '../../../table/components/cell';
 import {
     TableColumnSortOperation,
@@ -162,6 +166,12 @@ export class ColumnInternals<
      */
     @observable
     public hideHeaderIndicators = false;
+
+    /**
+     * How to align the header content.
+     */
+    @observable
+    public headerAlignment: TableColumnAlignment = TableColumnAlignment.left;
 
     /**
      * @internal Do not write to this value directly. It is used by the Table in order to store