+++ packages/nimble-components/src/table/index.ts
@@ -28,7 +28,7 @@ import {
     ExpandedState as TanStackExpandedState,
     OnChangeFn as TanStackOnChangeFn
 } from '@tanstack/table-core';
-import { keyEnter, keyShift } from '@microsoft/fast-web-utilities';
+import { keyEnter } from '@microsoft/fast-web-utilities';
 import { TableColumn } from '../table-column/base';
 import { TableValidator } from './models/table-validator';
 import { styles } from './styles';
@@ -235,7 +235,7 @@ export class Table<
     public tableScrollableMinWidth = 0;
 
     @observable
-    public documentShiftKeyDown = false;
+    public windowShiftKeyDown = false;
 
     private readonly table: TanStackTable<TableNode<TData>>;
     private options: TanStackTableOptionsResolved<TableNode<TData>>;
@@ -348,8 +348,9 @@ export class Table<
             passive: true
         });
         this.keyboardNavigationManager.connect();
-        document.addEventListener('keydown', this.onKeyDown);
-        document.addEventListener('keyup', this.onKeyUp);
+        window.addEventListener('keydown', this.onKeyDown);
+        window.addEventListener('keyup', this.onKeyUp);
+        window.addEventListener('blur', this.onBlur);
     }
 
     public override disconnectedCallback(): void {
@@ -357,8 +358,9 @@ export class Table<
         this.virtualizer.disconnect();
         this.keyboardNavigationManager.disconnect();
         this.viewport.removeEventListener('scroll', this.onViewPortScroll);
-        document.removeEventListener('keydown', this.onKeyDown);
-        document.removeEventListener('keyup', this.onKeyUp);
+        window.removeEventListener('keydown', this.onKeyDown);
+        window.removeEventListener('keyup', this.onKeyUp);
+        window.removeEventListener('blur', this.onBlur);
     }
 
     public checkValidity(): boolean {
@@ -405,7 +407,7 @@ export class Table<
         const selectionChanged = this.selectionManager.handleRowSelectionToggle(
             this.tableData[rowIndex],
             event.detail.newState,
-            this.documentShiftKeyDown
+            this.windowShiftKeyDown
         );
 
         if (selectionChanged) {
@@ -813,15 +815,15 @@ export class Table<
     };
 
     private readonly onKeyDown = (event: KeyboardEvent): void => {
-        if (event.key === keyShift) {
-            this.documentShiftKeyDown = true;
-        }
+        this.windowShiftKeyDown = event.shiftKey;
     };
 
     private readonly onKeyUp = (event: KeyboardEvent): void => {
-        if (event.key === keyShift) {
-            this.documentShiftKeyDown = false;
-        }
+        this.windowShiftKeyDown = event.shiftKey;
+    };
+
+    private readonly onBlur = (): void => {
+        this.windowShiftKeyDown = false;
     };
 
     private removeColumnObservers(): void {